better example of types use

+++ include/config.h
@@ -4,32 +4,36 @@
 #include "sqrt.h"
 // #include <math.h>
 
-// type for the accelerometer samples
-typedef int32_t accel_t;
+// type for the accelerometer samples, this depends on your hardware
+// for example, if your accelerometer uses 12bits then an int16_t would suffice
+typedef int16_t accel_t;
+
 // type used for time: warning the algorithm is not robust to roll-over of this variable
 // reset the algorithm if time overflows
 // preferably, time should be in ms, if not, use timeScalingFactor in the preprocessing stage to adjust it to ms
-typedef int64_t time_t;
+// example: a year worth of ms needs 35 bits, 32 bits allows you to store about 50 days of ms
+// as time doesn't go back, using unsigned integers makes sense
+typedef int32_t time_t;
 
 // magnitude of acceleration
 // this should be 1 bit bigger than accel_t
 // example: if the accelerometer works at 12 bits, magnitude should be 13 bits, therefore choosing int16_t for both will do
-typedef int64_t magnitude_t;
+typedef uint16_t magnitude_t;
 
-// accumulator of magnitude, shuold be bigger than magnitude
+// accumulator of magnitude, should be bigger than magnitude
 // the size depends on the length of the filter in the filtering stage
-// example: accelerometer works at 12 bits, the filter has 8 taps -> accumualtor should be at least 12 + 3 bits -> using int16_t will work fine
-typedef int64_t accumulator_t;
+// example: accelerometer works at 12 bits, the filter has 8 taps -> accumulator should be at least 12 + 3 bits -> using int16_t will do
+typedef int32_t accumulator_t;
 
 // steps count type
 // the size of this depends on what is the maximum number of steps you are willing to show and store
 // example: is the step count is reset every 24h and one walks no more than 20k / day (reasonable) -> 16 bits should be enough
-typedef int32_t steps_t;
+typedef uint16_t steps_t;
 
-// uncomment this line to disable interpolation
+/* Uncomment this line to disable interpolation */
 #define STEP_INTERPOLATION_DISABLE
 
-// incomment to skip filtering step
+// skip filtering step
 #define SKIP_FILTER
 
 // use this to allow dumping each stage on file, useful for debugging
